tests/wpt/web-platform-tests/cookie-store/serviceworker_cookieStore_subscriptions_reset.tentative.https.html

<!doctype html>
<meta charset="utf-8">
<title>Async Cookies: reset cookie change subscription list</title>
<link rel="help" href="https://github.com/WICG/cookie-store">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/service-workers/service-worker/resources/test-helpers.sub.js">
</script>
<script src='resources/helpers.js'></script>
<script>
const worker_url = 'serviceworker_cookieStore_subscriptions_reset.sub.js';
const scope = '/cookie-store/';

async function getSubscriptionCount(serviceWorker) {
  return sendMessageOverChannel('', serviceWorker);
}

promise_test(async t => {
  const registration = await service_worker_unregister_and_register(
      t, worker_url, scope);
  t.add_cleanup(() => registration.unregister());
  await wait_for_state(t, registration.installing, 'activated');
  assert_equals(await getSubscriptionCount(registration.active), 1,
                'subscription count before unregistration');

  await registration.unregister();

  const new_registration = await navigator.serviceWorker.register(
      worker_url, { scope });
  t.add_cleanup(() => new_registration.unregister());
  await wait_for_state(t, new_registration.installing, 'activated');

  assert_equals(await getSubscriptionCount(new_registration.active), 1,
                'subscription count after unregistration');
}, `cookiechange subscription list resets across service worker
    unregistrations`);
</script>
